+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changlog for the Musings Social Network API
 
 1. Created initial folder structure & files
+2. Added User and Thought models